Changing file suffixes to .*pp for consistency, adding missing copyright info, fixing broken macro usage, etc.
