+++ test/src/emptyfile
@@ -3,6 +3,8 @@
  *
  * Produced at the Lawrence Livermore National Laboratory.
  *
+ * LLNL-CODE-689114
+ *
  * All rights reserved.
  *
  * For release details and restrictions, please see RAJA/LICENSE.
