+++ test/unit/cpu/test-forall.cxx
@@ -18,7 +18,7 @@
 using namespace RAJA;
 using namespace std;
 
-#include "Compare.hxx"
+#include "Compare.hpp"
 #include "buildIndexSet.hxx"
 
 template <typename ISET_POLICY_T>
