+++ test/unit/test-type-traits.cpp
@@ -1,0 +1,69 @@
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
+// Copyright (c) 2016, Lawrence Livermore National Security, LLC.
+//
+// Produced at the Lawrence Livermore National Laboratory
+//
+// LLNL-CODE-689114
+//
+// All rights reserved.
+//
+// This file is part of RAJA.
+//
+// For additional details, please also read RAJA/README.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are met:
+//
+// * Redistributions of source code must retain the above copyright notice,
+//   this list of conditions and the disclaimer below.
+//
+// * Redistributions in binary form must reproduce the above copyright notice,
+//   this list of conditions and the disclaimer (as noted below) in the
+//   documentation and/or other materials provided with the distribution.
+//
+// * Neither the name of the LLNS/LLNL nor the names of its contributors may
+//   be used to endorse or promote products derived from this software without
+//   specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+// AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
+// ARE DISCLAIMED. IN NO EVENT SHALL LAWRENCE LIVERMORE NATIONAL SECURITY,
+// LLC, THE U.S. DEPARTMENT OF ENERGY OR CONTRIBUTORS BE LIABLE FOR ANY
+// DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+// DAMAGES  (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS
+// OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+// HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
+// STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING
+// IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+// POSSIBILITY OF SUCH DAMAGE.
+//
+//~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
+
+#include "RAJA/RAJA.hpp"
+
+#include "RAJA/internal/type_traits.hpp"
+
+#include "gtest/gtest.h"
+
+#ifdef RAJA_ENABLE_OPENMP
+static_assert(!RAJA::is_sequential_policy<RAJA::omp_parallel_for_exec>::value, "");
+static_assert(RAJA::is_openmp_policy<RAJA::omp_parallel_for_exec>::value, "");
+static_assert(!RAJA::is_cuda_policy<RAJA::omp_parallel_for_exec>::value, "");
+#endif
+static_assert(RAJA::is_sequential_policy<RAJA::seq_exec>::value, "");
+static_assert(!RAJA::is_openmp_policy<RAJA::seq_exec>::value, "");
+static_assert(!RAJA::is_cuda_policy<RAJA::seq_exec>::value, "");
+#ifdef RAJA_ENABLE_CUDA
+static_assert(!RAJA::is_sequential_policy<RAJA::cuda_exec<128>>::value, "");
+static_assert(!RAJA::is_openmp_policy<RAJA::cuda_exec<128>::value, "");
+static_assert(RAJA::is_cuda_policy<RAJA::cuda_exec<128>>::value, "");
+static_assert(!RAJA::is_sequential_policy<RAJA::cuda_exec_async<128>>::value, "");
+static_assert(!RAJA::is_openmp_policy<RAJA::cuda_exec_async<128>::value, "");
+static_assert(RAJA::is_cuda_policy<RAJA::cuda_exec_async<128>>::value, "");
+#endif
+
+TEST(TypeTraits, Default)
+{
+  ASSERT_EQ(true, true);
+}
